requirements.txt
Pillow
requests

